pyproject.toml
[project]
name = "gemini-lk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[google]~=1.3",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "livekit-plugins-openai>=1.3.10",
    "python-dotenv>=1.2.1",
]
